+++ plugins/ingest/src/main/java/org/elasticsearch/plugin/ingest/transport/simulate/SimulatePipelineTransportAction.java
@@ -48,13 +48,12 @@ public class SimulatePipelineTransportAction extends HandledTransportAction<Simu
     protected void doExecute(SimulatePipelineRequest request, ActionListener<SimulatePipelineResponse> listener) {
         Map<String, Object> source = XContentHelper.convertToMap(request.getSource(), false).v2();
 
-        ParsedSimulateRequest simulateRequest;
-        ParsedSimulateRequest.Parser parser = new ParsedSimulateRequest.Parser();
+        SimulatePipelineRequest.Parsed simulateRequest;
         try {
             if (request.getId() != null) {
-                simulateRequest = parser.parseWithPipelineId(request.getId(), source, request.isVerbose(), pipelineStore);
+                simulateRequest = SimulatePipelineRequest.parseWithPipelineId(request.getId(), source, request.isVerbose(), pipelineStore);
             } else {
-                simulateRequest = parser.parse(source, request.isVerbose(), pipelineStore);
+                simulateRequest = SimulatePipelineRequest.parse(source, request.isVerbose(), pipelineStore);
             }
         } catch (IOException e) {
             listener.onFailure(e);
